+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Cleaned up some warnings and unnecessary logging
 - Fixed a bug with how different browsers handle 'indexeddb://' plugin urls - #2
+- Fixed a bug with how extra plugin lists are stored/updated - #3
 
 ## [0.6.0] - 2022-08-30